+++ apps/backoffice/src/lib/be/__tests__/services.test.ts
@@ -159,6 +159,14 @@ const {
   })
 }));
 
+const { isAxiosError } = vi.hoisted(() => ({
+  isAxiosError: vi.fn().mockReturnValue(false)
+}));
+
+vi.mock("axios", async () => ({
+  isAxiosError
+}));
+
 vi.mock("@/lib/be/subscription-migration-client", () => ({
   getSubscriptionsMigrationClient
 }));
@@ -461,7 +469,12 @@ describe("Services TEST", () => {
 
       const result = await retrieveServiceList(mocks.anUserId, 10, 0);
 
-      expect(getServiceListMock).toHaveBeenCalledWith(mocks.anUserId, 10, 0, undefined);
+      expect(getServiceListMock).toHaveBeenCalledWith(
+        mocks.anUserId,
+        10,
+        0,
+        undefined
+      );
       expect(bulkFetchLifecycleMock).toHaveBeenCalledWith([
         aServiceinPublicationId,
         aServiceNotInPublicationId
@@ -514,13 +527,61 @@ describe("Services TEST", () => {
 
       const result = await retrieveServiceList(mocks.anUserId, 10, 90);
 
-      expect(getServiceListMock).toHaveBeenCalledWith(mocks.anUserId, 10, 90, undefined);
+      expect(getServiceListMock).toHaveBeenCalledWith(
+        mocks.anUserId,
+        10,
+        90,
+        undefined
+      );
       expect(bulkFetchLifecycleMock).not.toHaveBeenCalled();
       expect(bulkFetchPublicationMock).not.toHaveBeenCalled();
 
       expect(result).toStrictEqual({
         value: [],
         pagination: { count: 90, limit: 10, offset: 90 }
+      });
+    });
+
+    it("when APIM returns 404 neither service-lifecycle or service-publication bulkFetch method should be called", async () => {
+      const bulkFetchLifecycleMock = vi.fn(() => TE.right([]));
+      const bulkFetchPublicationMock = vi.fn(() => TE.right([]));
+
+      const getServiceListMock = vi.fn(() =>
+        TE.left({
+          message: "Received 404 response",
+          response: { status: 404 }
+        })
+      );
+
+      getServiceLifecycleCosmosStore.mockReturnValueOnce({
+        bulkFetch: bulkFetchLifecycleMock
+      });
+      getServicePublicationCosmosStore.mockReturnValueOnce({
+        bulkFetch: bulkFetchPublicationMock
+      });
+
+      getApimRestClient.mockReturnValueOnce(
+        Promise.resolve({
+          getServiceList: getServiceListMock
+        })
+      );
+
+      isAxiosError.mockReturnValueOnce(true);
+
+      const result = await retrieveServiceList(mocks.anUserId, 10, 90);
+
+      expect(getServiceListMock).toHaveBeenCalledWith(
+        mocks.anUserId,
+        10,
+        90,
+        undefined
+      );
+      expect(bulkFetchLifecycleMock).not.toHaveBeenCalled();
+      expect(bulkFetchPublicationMock).not.toHaveBeenCalled();
+
+      expect(result).toStrictEqual({
+        value: [],
+        pagination: { count: 0, limit: 10, offset: 90 }
       });
     });
 
@@ -579,7 +640,12 @@ describe("Services TEST", () => {
 
       const result = await retrieveServiceList(mocks.anUserId, 10, 0);
 
-      expect(getServiceListMock).toHaveBeenCalledWith(mocks.anUserId, 10, 0, undefined);
+      expect(getServiceListMock).toHaveBeenCalledWith(
+        mocks.anUserId,
+        10,
+        0,
+        undefined
+      );
       expect(bulkFetchLifecycleMock).toHaveBeenCalledWith([
         aServiceInLifecycleId,
         aServiceNotInLifecycleId
@@ -639,7 +705,12 @@ describe("Services TEST", () => {
         retrieveServiceList(mocks.anUserId, 10, 0)
       ).rejects.toThrowError();
 
-      expect(getServiceListMock).toHaveBeenCalledWith(mocks.anUserId, 10, 0, undefined);
+      expect(getServiceListMock).toHaveBeenCalledWith(
+        mocks.anUserId,
+        10,
+        0,
+        undefined
+      );
       expect(bulkFetchLifecycleMock).toHaveBeenCalledWith([
         mocks.aBaseServiceLifecycle.id
       ]);
@@ -685,7 +756,12 @@ describe("Services TEST", () => {
         retrieveServiceList(mocks.anUserId, 10, 0)
       ).rejects.toThrowError();
 
-      expect(getServiceListMock).toHaveBeenCalledWith(mocks.anUserId, 10, 0, undefined);
+      expect(getServiceListMock).toHaveBeenCalledWith(
+        mocks.anUserId,
+        10,
+        0,
+        undefined
+      );
       expect(bulkFetchLifecycleMock).toHaveBeenCalledWith([
         mocks.aBaseServiceLifecycle.id
       ]);
@@ -720,7 +796,12 @@ describe("Services TEST", () => {
         retrieveServiceList(mocks.anUserId, 10, 0)
       ).rejects.toThrowError();
 
-      expect(getServiceListMock).toHaveBeenCalledWith(mocks.anUserId, 10, 0, undefined);
+      expect(getServiceListMock).toHaveBeenCalledWith(
+        mocks.anUserId,
+        10,
+        0,
+        undefined
+      );
       expect(bulkFetchLifecycleMock).not.toHaveBeenCalled();
       expect(bulkFetchPublicationMock).not.toHaveBeenCalled();
     });