+++ apps/backoffice/CHANGELOG.md
@@ -1,5 +1,17 @@
 # io-services-cms-backoffice
 
+## 0.15.3
+
+### Patch Changes
+
+- 0f56282: ock services-cms /info only in development environment
+
+## 0.15.2
+
+### Patch Changes
+
+- 83059a6: Handle 404 Apim getServiceList response
+
 ## 0.15.1
 
 ### Patch Changes